Add environment variable to disable clean on inotify
When using ampache on external storage that doesn't have a guaranteed
availability, the inotifywait monitor will clean the catalogue if the
external storage goes down. This adds a new environment variable to docker
that switches the monitor from running -cage to running -age instead.

That way, if the external storage is unavailable, the catalogue is preserved
rather than automatically deleted. (-age doesn't remove files if they don't
exist).